tests/wpt/tests/css/css-color/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Color properties</title>
<link rel="help" href="https://drafts.csswg.org/css-color/#property-index">
<meta name="assert" content="Properties inherit or not according to the spec.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>
assert_inherited('color', 'rgb(0, 0, 0)', 'rgb(42, 53, 64)');
assert_not_inherited('opacity', '1', '0.5');
</script>
</body>
</html>
